app/bibinet_fastapi/requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
asyncpg==0.29.0
python-dotenv==1.0.1
